Decoupled two types of templates from each other.
This removes auto-generation of tags for the default (fmt) templates,
which was causing issues with nil/empty template variables, as when the
values are not provided, the check `if param` fails, as `param` gets
assigned a table structure to support tag generated. This processing now
is limited to only the table-based template system.